+++ src/wamp_client_app.erl
@@ -55,10 +55,11 @@ stop(_State) ->
 
 
 % tracer() ->
-%     dbg:tracer(),
-%     dbg:p(all, c),
-%     dbg:tpl(awre_sup, '_', x),
-%     dbg:tpl(awre_con, '_', x),
-%     % dbg:tpl(awre_trans_tcp, '_', x).
-%     dbg:tpl(wamp_client_peer_sup, '_', x),
-%     dbg:tpl(wamp_client_peer, '_', x).
+    % dbg:tracer(),
+    % dbg:p(all, c),
+    % dbg:tpl(awre_sup, '_', x),
+    % dbg:tpl(awre_con, '_', x),
+    % dbg:tpl(wamp_client_sensitive, '_', x),
+    % dbg:tpl(awre_trans_tcp, '_', x).
+    % dbg:tpl(wamp_client_peer_sup, '_', x),
+    % dbg:tpl(wamp_client_peer, '_', x).